feat: Add Working Day 13 solution + tests
